requirements.txt
streamlit>=1.24.0
Pillow>=9.5.0
numpy>=1.24.3
pytest>=7.3.1
streamlit-components-bootstrap==0.1.2
streamlit-materialize==0.1.0
streamlit-custom-css==0.1.0
requests>=2.28.0